(PC-19757) refactor(share): extract useOfferUrl in helpers folder
