+++ src/features/share/helpers/getOfferUrl.ts
@@ -1,0 +1,7 @@
+import { getScreenPath } from 'features/navigation/RootNavigator/linking/getScreenPath'
+import { WEBAPP_V2_URL } from 'libs/environment'
+
+export function getOfferUrl(id: number): string {
+  const path = getScreenPath('Offer', { id })
+  return `${WEBAPP_V2_URL}${path}`
+}
